app/requirements.txt
Flask==2.3.3
torchvision==0.16.2
torchdata==0.7.1
torchtext==0.16.2
portalocker==2.8.2
datasets==2.16.1
spacy==3.7.1
matplotlib==3.8.2
numpy==1.26.3
scipy==1.12.0
